feat: The added view presents the achieved awards for the user
This view shows the awards that will be a gray color when the awards are not achieved with the ability to tap each icon and reveal what the user must do to get the award. The award wil turn to the color provided by the Awards.json file when the user meets the criterion for that award. This view was made by adding a foreach loop with that cycles through the different awards and presents them in a lazy grid. An alert is displayed on tap of the button to show the criteria needed to achieve the specific award that is pressed