feat: The added view displays on the homescreen when the user adds items to projects to be completed
This view uses all items that are in an open project to display each item with the ability to edit the item upon a press. The Progress ring next to each item title shows the prcentage complete marked with an icon. A custom blur view modifer was added to the background color of each cell the item is in. When no projects are selected an empty view will present until items are added.The progress bar will display in the color that matches the color selected when creating a new project to give the user more context