faet: This added view gives the user the option to Edit the current Item that is selected
This view presents through a Navigation view and gives the ability to set the Item title, description, priority (which will change where the item displays in the itemlist) and if the item is complete. The edited item then syncs to Core Data upon closing of the app or when the user exits the navigation view. The priority is written through the use of a custom onchange modifer that stores the tag of each of the priority being shown in order to be sorted.