feat: The added view gives the user the ability to edit a project
The User is able to enter a name for the project, project color (which will display in the accent color throughout), description of the project, or the ability to close the project. When the project is closed it will move into the Closed Projects tab and the items in that project will not display on the homescreen. The user may also delete a project that prompts the user with a warning to make sure they want to delete the project as it will erase any items in the project as well. The data entered is saved into Core Data and can be saved through the user losing focus of the application or through exiting the navigation link.