feat: The added view displays all items within the header of their project
This gives the user the ability to select and edit all the items in the given project that are organized by each project. The user can select whether they want the list to prsent in order of creation date, completion, or optimized which takes into account several factors that make the most relevent items show at the top of the list. The items have an icon next to them to show their completion through a check marked circle when the item is complete in the projects accent color, an exclimation point if the project have a high priority, and no icon when the item is neither complete nor a high priority.