feat: Added this view to display ProjectHeadersView, Item Lists, navigation to editing projects and items, and addition of new projects
This view is seperated into filtering open projects and closed projects. A new project is able to be added with the plus button that is in the trailing position of the navigation view. In the leading position of the navigation bar there is crossed arrows that allow the user to pick the filter for the list items for optamized, creation date, or completion amount. The add item button at the bottom of the item list allows the user to add a new item that can be edited upon press of the given item to bring up the ItemEditView. The user can swipe to delete an item in the item list. The views to filter open and closed projects are done so using tags that are called in the ContentView. All edits made are saved to the dataController which in turn saves to Core Data so the projects edit and item added will be saved