refactor: Added MVVM architecture
This refactoring took place by adding a ProjectsViewModel and abstracting the Code responsible for the Data prcessing out of the ProjectsView. The remaining code strictly works with the UI elements. The refactoring was done to reflect a popular use of MVVM making the views, view models, models sperated. The data is communicated through Publishing the changes throughout to Bind the data to the Views from the ViewModels.