refactor: refactored to MVVM architecture
Refactored ItemRow to pull out controllers into seperate viewmodel. This was done by seperating the icon and color from the the View and conforming to the Observable Object Protocol in order for the ItemRowView to Observe the changes made. This refacotring helps conform to the MVVM architecture and keeps the UI elements in the view while placing the controllers in a seperate file for easier access and understanding of how the Item row is being manipulated by the user as well as the UI elements