refactor: Refactored to MVVM architecture
This refactoring took place by abstracting the Controlling elements of the Data that is related to Core Data into a seperate file while leaving the UI elements in the view to be called by an Observable Object. Both fetch requests were made for items and projects through the added itemsController and projectController. The view model conforms to  NSObject, ObservableObject, NSFetchedResultsControllerDelegate in order to integerate Core Data releaince while also being able to be referenced in the UI Homeview as Observed Objects to call the view model into the view to store, present, and save changes made by the user. This refactoring took place in order to conform to MVVM as to better understanding the controlling elements of the view as well as the UI elements by seperating each into a viewModel and View respectivley