fix: Fixed bug
The bug fixed in this commit deals with the inability to close a project and reopen it when the project is closed. This bug was fixed by changing the project varible in EditProjects to an Observable Object so that the different views are always communicating with each other upon change. The Tests were rerun succesfully to present this change