feat: Added weekDays static Varible
The added varibles store the added String for day into Core Data. This is done to persist the data for the selected day and attach it to the Project. This allows the Recipe to persist on a selected day of the week that the meal is able to be planned each week as shown in the MealsThisWeekView