style: Added animation to check off button
The added animation has two parts: On press of the icon or label that completes the item with a custom feedback generation for complete and unchecked. The icon displays a ring around the cicle and the circle shinks as the user continues to press until the item is completed upon release. The item title was placed in a HStack with an amount placed in perentheses that can be added in the project edit view