feat: Added Settings menu to select colors for application
This view was added in order to let the user choose a color from the assests catalog for background color of the circles in the Home Tab, foreground color of the smaller circles in the Home Tab, and also the Accent color throughout the application. In order to add persistance after the user chooses a color, User defaults are used by an  @objc dynamic varible as an extension of the view, to get and set the values which are then published through an observable object class. The color picker uses the added colors as Strings to be saved to the user defaults and called in the HomebackgroundView