feat: Added List View for Purchased Items
When an Ingredient is checked off in the shopping list it will still persist but will show up in the Purchased tab to give the user more context of what Items are already bought without having to look through the Shopping List. The Items are presented in a list with a background of  a rounded rectangle with a border gradiant of the custom Recipe color the user chooses.