feat: Cleaned up menu for EditProjectView
Removed unneccasry Move to meals button in Form. This was done becuase the item is autmatically added to meals this week upon exit to the unassigned category. Also removed the move to Purchase button as the Purchased view was refactored to only show the compelted items. The two buttons left are clear and delete which either clears the recipe from the Shopping List without effecting Saved views or Meals this week and the delete button which removes the recipe entirley